star tool should be fixed; needs testing
