+++ src/main/java/com/alysaa/geyseradmintools/Gat.java
@@ -1,17 +1,17 @@
 package com.alysaa.geyseradmintools;
 
 import com.alysaa.geyseradmintools.commands.GatCommand;
-import com.alysaa.geyseradmintools.listeners.*;
+import com.alysaa.geyseradmintools.listeners.AdminLockChat;
+import com.alysaa.geyseradmintools.listeners.AdminToolInventory;
+import com.alysaa.geyseradmintools.listeners.AdminToolOnDeath;
+import com.alysaa.geyseradmintools.listeners.AdminToolOnJoin;
+import com.alysaa.geyseradmintools.listeners.AdminToolOnRespawn;
+import com.alysaa.geyseradmintools.utils.ItemStackFactory;
 import com.alysaa.geyseradmintools.utils.bstats.Metrics;
 import org.bukkit.Bukkit;
-import org.bukkit.ChatColor;
-import org.bukkit.Material;
 import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
-import org.bukkit.entity.Item;
-import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import java.io.File;
@@ -23,8 +23,6 @@ public class Gat extends JavaPlugin {
     public static Gat plugin;
     public static Logger logger;
 
-    private static ItemStack starTool;
-
     @Override
     public void onEnable(){
         new Metrics(this, 10943);
@@ -33,19 +31,13 @@ public class Gat extends JavaPlugin {
 
         createFiles();
         checkConfigVer();
-
+        ItemStackFactory.createStarTool();
         this.getCommand("gadmin").setExecutor(new GatCommand());
         Bukkit.getServer().getPluginManager().registerEvents(new AdminToolOnJoin(), this);
         Bukkit.getServer().getPluginManager().registerEvents(new AdminLockChat(), this);
         Bukkit.getServer().getPluginManager().registerEvents(new AdminToolInventory(), this);
         Bukkit.getServer().getPluginManager().registerEvents(new AdminToolOnRespawn(), this);
         Bukkit.getServer().getPluginManager().registerEvents(new AdminToolOnDeath(), this);
-
-        // Create our item tool to open the main form
-        starTool = new ItemStack(Material.NETHER_STAR, 1);
-        ItemMeta starToolMeta = starTool.getItemMeta();
-        starToolMeta.setDisplayName(ChatColor.translateAlternateColorCodes('&', "&6Admin Tools"));
-        starTool.setItemMeta(starToolMeta);
 
         getLogger().info("Plugin has been enabled - Provided by ProjectG");
 
@@ -73,14 +65,5 @@ public class Gat extends JavaPlugin {
         } catch (IOException | InvalidConfigurationException e) {
             e.printStackTrace();
         }
-    }
-
-    /**
-     * Get the star tool as an ItemStack object
-     *
-     * @return the ItemStack
-     */
-    public static ItemStack getStarTool() {
-        return starTool;
     }
 }
