+++ src/main/java/com/alysaa/geyseradmintools/listeners/AdminToolOnRespawn.java
@@ -2,19 +2,18 @@ package com.alysaa.geyseradmintools.listeners;
 
 import com.alysaa.geyseradmintools.Gat;
 import com.alysaa.geyseradmintools.utils.CheckJavaOrFloodPlayer;
+import com.alysaa.geyseradmintools.utils.ItemStackFactory;
 import org.bukkit.ChatColor;
-import org.bukkit.Material;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerRespawnEvent;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.meta.ItemMeta;
 
 public class AdminToolOnRespawn implements Listener {
 
-    private final ItemStack starTool = Gat.getStarTool();
+    private static final ItemStack starTool = ItemStackFactory.getStarTool();
 
     @EventHandler
     public void onRespawn(PlayerRespawnEvent e) {
